requirements.txt
tensorflow-gpu
numpy
mtcnn
opencv-python
